+++ src/mt/_core/__init__.py
@@ -1,15 +1,13 @@
 """Core
 ---
 * settingに必要な仕組みの提供
-* 具体的な理論のためのデータ構造・仕組みは扱わない ⇒ shared
 """
 
 # utils
 # interfaces
 from . import context, interfaces, utils
-
 # implementation
-from .object import ObjectMT
-from .setting import SettingMT
+from .object import Object
+from .setting import Setting
 
-__all__ = ["ObjectMT", "SettingMT", "interfaces", "context", "utils"]
+__all__ = ["Object", "Setting", "interfaces", "context", "utils"]
