+++ src/mt/_core/context/_inject.py
@@ -12,7 +12,7 @@ def inject[**P, R](func: t.Callable[P, R]) -> t.Callable[P, R]:
         from ...utils import Context
 
         if "setting" not in kwargs.keys():
-            setting = Context.setting.get(None)
+            setting = Context.internal.get(None)
             return func(*args, setting=setting, **kwargs)
         return func(*args, **kwargs)
