--- src/mt/_core/settings/__init__.py
@@ -1,3 +1,0 @@
-from .parsedsetting import ParsedSetting
-
-__all__ = ["ParsedSetting"]
