+++ src/mt/_core/shared/__init__.py
@@ -3,5 +3,4 @@
 * 具体的な理論のためのデータ構造・仕組みを提供
 """
 
-from .harmonicsystem import HarmonicSystem
 from .intervals import Intervals
