+++ src/mt/note/note.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import typing as t
 from typing import TYPE_CHECKING
 
-from .._core import ObjectMT
+from .._core import Object
 from .._core.context import inject
 from ._base import BaseNote
 
@@ -11,7 +11,7 @@ if TYPE_CHECKING:
     from scale import Scale
 
 
-class Note(BaseNote, ObjectMT):
+class Note(BaseNote, Object):
 
     @inject
     def __init__(
@@ -20,7 +20,7 @@ class Note(BaseNote, ObjectMT):
         super().__init__(**kwargs)
 
     def is_notenumber(self, notenumber: t.Any) -> t.TypeGuard[int]:
-        return self.setting.notenumber.is_notenumber(notenumber)
+        return self.scheme.is_notenumber(notenumber)
 
     def is_notename(self, value: t.Any) -> t.TypeGuard[str]:
         return
