+++ src/mt/scale/scale.py
@@ -2,7 +2,7 @@ import typing as t
 from functools import cached_property
 from itertools import accumulate, compress
 
-from .._core import ObjectMT
+from .._core import Object
 from .._core.context import inject
 from .._core.utils import classproperty, rotate
 from ..note import Key, PitchClass
@@ -10,7 +10,7 @@ from ._base import BaseScale
 from .processing.accidentals import accidentals
 
 
-class Scale(BaseScale, ObjectMT):
+class Scale(BaseScale, Object):
 
     _intervals: t.ClassVar[tuple[int]]
