+++ src/setting.toml
@@ -1,4 +1,4 @@
-[scheme]
+[pitchclass]
 semitone  = 12
 intervals = [2,1,2,2,1,2,2]
 symbols   = ["A","B","C","D","E","F","G"]
@@ -7,18 +7,18 @@ accidental.upper_symbol = "#"
 accidental.lower_symbol = "b"
 
 
-[notenumber]
+[note]
 min = 0
 max = 128
 
-[[notenumber.presentations]]
+[[note.presentations]]
 name = "scientific pitch notation"
 areas = "SPN"
 symbols = ["<N>-1","<N>0","<N>1","<N>2","<N>3","<N>4","<N>5","<N>6","<N>7","<N>8","<N>9"]
 reference.number = 69
 reference.symbol = 6
 
-[[notenumber.presentations]]
+[[note.presentations]]
 name = "helmholtz pitch notation"
 areas = "HPN"
 symbols = ["<N>,,,","<N>,,","<N>,","<N>","<n>'","<n>''","<n>'''","<n>''''","<n>'''''","<n>''''''","<n>'''''''"]
